model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
denoising-diffusion-pytorch>=2.0.0
ema-pytorch>=0.5.0
Pillow>=9.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
scipy>=1.10.0
accelerate>=0.20.0
